+++ src/main/java/de/nordakademie/iaa_multiple_choice/web/StudentSubmitAnswerAction.java
@@ -59,7 +59,7 @@ public class StudentSubmitAnswerAction extends BaseStudentExamAction {
             throw new QuestionNotFoundException();
         } else {
             final Optional<Question> optionalQuestion = getExam().getQuestions().stream()
-                    .filter(q -> getQuestionId().longValue() == q.getId().longValue()).findFirst();
+                    .filter(q -> getQuestionId() == q.getId()).findFirst();
             if (!optionalQuestion.isPresent()) {
                 throw new QuestionNotFoundException();
             }
